fix 修正批量重启 redis 进程的脚本实现逻辑

+++ compose_local_redis.py
@@ -21,32 +21,34 @@ def up():
   if os.path.exists(gidfile):
     click.echo(f'{gidfile} already exists', err=True)
     _try_kill_exists_process_group()
-  click.echo(f'before gid {os.getgid()} pid={os.getpid()} ppid={os.getppid()}, pgrp={os.getpgrp()}')
-  os.setpgrp()
-  click.echo(f'after gid {os.getgid()}')
-  gid = os.getppid()
-  with codecs.open(gidfile,mode='wt') as f:
-    f.write(str(gid))
-
   cwd = os.path.dirname(__file__)
+  proc_ids = []
   for port in range(6380,6383):
     conf = f"conf.d/redis-{port}.conf"
     proc = subprocess.Popen(f'redis-server {conf}',cwd=cwd, shell=True)
+    proc_ids.append(proc.pid)
   for port in range(26380,26383):
     conf = f"conf.d/sentinel-{port}.conf"
     proc = subprocess.Popen(f'redis-sentinel {conf}',cwd=cwd, shell=True)
+    proc_ids.append(proc.pid)
+
+  with codecs.open(gidfile,mode='wt') as f:
+    f.writelines([str(pid) for pid in proc_ids])
 
 def _try_kill_exists_process_group():
   if not os.path.exists(gidfile):
     click.echo(f'{gidfile} not found', err=True)
     return
   with codecs.open(gidfile,mode='rt') as f:
-    gid = int(f.read().strip())
-  click.echo(f"Terming {gid}")
-  try:
-    os.kill(gid,signal.SIGTERM)
-  except Exception as e:
-    click.echo(message=str(e),err=True)
+    for line in f.readlines():
+      if not line:
+        continue
+      gid = int(line.strip())
+      click.echo(f"Terming {gid}")
+      try:
+        os.kill(gid,signal.SIGTERM)
+      except Exception as e:
+        click.echo(message=str(e),err=True)
   os.remove(gidfile)
 
 @cli.command()
